schema/snapcraft.json: allow type: os with build-base
This is to build the core snap with a modern version of snapcraft.

We specifically only allow type: os with build-base required and not allowed
with base set, because the only type: os snap that should ever be built is the
core snap, which will not set a base because it is a base.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>