meta: add suite/component validation checks for package-repositories
According to apt's sources.list manpage, there are two invalid cases
that snapcraft should verify:

(1) No components must be specified if suite is a path.  Apt determines
that a suite is a path if it ends with a "/".

(2) Due to the above requirement, it follows that suite types cannot
be mixed in the same repository object.

This commit makes the following changes:

- Add these checks to package-repositories when being unmarshalled.

- Note #1 in the specification and JSON schema.

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>