build providers: unified provider refactoring for provider setup
Refactoring for unified approach to checking if a provider is ready, and
performing installation.

- Adds 'is_provider_ready()' interface, replacing 'ensure_provider()'.

- Makes 'setup_provider()' responsible for all user interaction to install
and/or configure the provider, and its depdenencies.  Adds 'interactive'
parameter to allow for non-interactive installs, though not currently utilized.

- Refactored LXD checks into their own methods, adding a compatibility check to
check the LXD version is 4.0+, and removing the snap install check in favor of
just ensuring the hardcoded /snap/bin/lx[cd] paths are present.

- Moves setup code out of MultipassCommand into the Multipass provider.

- Update lifecycle CLI to utilize new interfaces.  When tty is not connected,
and setup is required, exit with error that that build provider is not
installed or configured.

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>