cli: support snap --output <directory>
Currently one may only specify an explicit path to output snap
in the case of destructive mode.  For LXD/Multipass, one may only
specify the name of the snap within the project directory, otherwise
it gets "lost" in the container and is not retrieved.

This change allows output to point to an output directory, rather
than a snap file name.  This will help providers that may ship
multiple snaps, or users that want the snapcraft-provided name,
but rather have it shipped into a different directory.

NOTE that this change does not fix the still-broken case where
LXD/Multipass users specify a path outside of the project. This
should be addressed in upcoming work.

- Update _pack() to support the three uses of 'output'
- Introduce test coverage for _pack()

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>