pluginhandler: support using patchelf on strict snaps (#3277)
Patchelf is automatically used on classic snaps to patch the rpath and
interpreter of ELF files in the snap to find things in the snap itself
without requiring LD_LIRARY_PATH to be set, etc. This is usually
desired, but in the cases that it's not, the snapcraft CLI supports a
`no-patchelf` build attribute to disable this behavior.

There are situations where strictly-confined snaps could benefit from
this behavior as well (e.g. content sharing, plugin models, etc.). Add a
new build attribute called `enable-patchelf` to opt in to rpath and
interpreter patching in a strictly-confined snap.

Signed-off-by: Kyle Fazzari <kyrofa@ubuntu.com>