storeapi: drop arch requirement for get_channel_mapping()
The base build provider and SnapInjector() specifies the architecture of the
snap to download/install, which causes unnecessary coupling of the host <->
build provider architecture. While this could trivially determined, it appears
to be unnecessary as Snapd assumes the correct architecture to match the runtime
environment.

To preserve the functionality for potential usage in the future, this change
makes the arch field optional in get_channel_mapping() and download() store
APIs.

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>