snapcraftctl: add checks for empty string for set-version & set-grade
It's a pretty common mistake for scriplets to end up setting "" in
set-version, leading the the following error:
`Failed to generate snap metadata: 'adopt-info' refers to part <part>,
but that part is lacking the 'parse-info' property.`

Setting an empty grade effectively just ignores the invalid value,
relying on the default grade later.  Check this for "stable" or
"devel".

Explicitly error out with helpful messages when invalid options
are used.

- Add unit tests for coverage.

- Update related tests that set invalid grades.

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>