pluginhandler: properly handle snapcraftctl errors (#3317)
Address issues including:

- snapcraftctl operations always exiting with 0, even for failure
  cases.

- snapcraft override scripts continue to run even after
  parent snapcraft process fails.  This is because `set -e` does
  not catch the error when snapcraftctl returns 0, and nothing
  is otherwise done to terminate the script.

Given the nature of these symptoms, this is likely the cause of
a number of other seemingly unrelated issues.

These issues occur because _handle_builtin_function() only
catches exceptions of type errors.ScriptletBaseError,
ignoring other cases that may result from the build/pull operations,
such as NodejsPluginMissingPackageJsonError, or
SnapcraftPluginBuildError.

Instead of trying to manage exceptions in _handle_builtin_function(),
catch _any_ unhandled exception in _run_scriplet()'s try block. This
simplifies the more complex return value and feedback handling.

Don't print an error from the feedback channel in snapcraftctl,
the parent process will take care of it.  This is necessary
because the error that is now seen by snapcraftctl would lead
to redundant output.  Just exit with -1 when an error is
read from the feedback fifo.

LP: #1898922

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>