cli: warn when running snapcraft inside assets directory
Because snapcraft will use a snapcraft.yaml in the current directory,
snapcraft may be run inside the assets directory instead of the
expected project directory.  This can be a non-obvious error for the
user.

See:
https://forum.snapcraft.io/t/help-with-desktop-file-not-being-primed/21111

This adds a warning if the user is running inside a directory named
"snap" or "build-aux".

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>