multipass build provider: check if instance exists before deleting
Snapcraft may attempt to delete a multipass instance that does
not exist. Check to see if instance exists prior to calling delete.

This may occur if Snapcraft sees project provider directory, but
the instance does not exist when it goes to clean it.

This commit adds a new exists() method to multipass commands. While
it's not a "real" multipass command, it seems to be a fair interface
to provide without overloading the usage of info() or another command.

LP #1904774

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>