project loader: advanced grammar support for lists
The grammar expects that it always works with a list of strings,
and treats those string primitives as a set.

This works for all of the current grammar usage which acts on
lists that are intended to work as sets.  But it will fail when
expanded to work on other cases where ordering matters (e.g.
build-environment), or when primtives are not a string (e.g.
build-environment objects).

This is addressed by:

- Converting all grammars' set usage into lists.

- Updating GrammarProcessor to allow it to parse dictionaries
  and treat unhandled section dictionaries as primitives.

- Update the part grammar to convert the returned lists into
  sets for the currently used properties.

- Immediately process finalized statements to ensure ordering
  is consistent (rather than processing all statements at the
  end).  Remove StatementCollection.process_all() now that it
  is unused.

Update tests accordingly.

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>