grammar: fix the case of disappearing statements
Any statement prior to an 'on.. to..' clause vanishes into
the abyss.

E.g.:

```
stage-packages:
  - on amd64 to arm64:
    - foo
  - on amd64 to i386:
    - foo2
```

In this case `foo` will never be added to the package list.

When starting to process an on/to statement, ensure the previous
one is added to the statement collection.

Add a test to cover this case and update the basic grammar tests
to explicitly declare the target (deb) arch.

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>