repo: default to target arch for stage package cache
- Add interface to project to define which arch should
  be used for staging packages.

- Add target_arch parameter to fetch_stage_packages().

- Update apt_cache to support specifying the default
  architecture for the stage package cache.

If running with an existing cache, check if <cache>/etc/apt
is a symlink (how it was done up until now) and unlink it.
Then copy the /etc/apt tree instead of linking it. Finally,
add a config file to specify the default target architecture.
Write that config file out every time the stage cache is
(re)initialized.

- Update all usage of fetch_stage_packages() to comply
  with new interface.

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>