snap: explicitly add setuptools-rust package to python-packages
With cryptography >= 3.4, setuptools-rust is required to build.

It is listed as part of the setup_requires [1], but this is not
accounted for in the v1 python plugin.  This is due to a discrepancy
of dependency resolution of pip download vs install with regards
to setup_requires [2].

[1] https://github.com/pyca/cryptography/blob/main/setup.py#L46
[2] https://github.com/pypa/pip/issues/5865

Work around this by simply pinning the requirement in the
snapcraft.yaml parts (as it is a Snapcraft-related issue).

Additionally add rustc to build-packages to ensure it is available
if the package is built from source.

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>