python v2 plugin: fix multiple python parts with staged python
Fix a regression where snaps with multiple python parts and staged
python packages may cause a file conflict & broken (unused) symlink.

When creating a venv with "/usr/bin/python3", the venv will populate
symlinks for:
/bin/python
/bin/python3

When creating a venv with "/usr/bin/python3.8", the venv will populate
symlinks for:
/bin/python
/bin/python3
/bin/python3.8

By resolving the interpreter prior to using it, the venvs were getting
an additional symlink which are not being fixed up like "python3",
or $SNAPCRAFT_PYTHON_INTERPRETER.

Should SNAPCRAFT_PYTHON_INTERPRETER be set to a more specific minor
version, either the snap must filter out the additional symlink or the
plugin needs to be updated to correct, or remove, the problematic
links.

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>