meta: add support for `kernel.yaml` for kernel snaps (#3464)
Snapd supports the "kernel.yaml" file for kernel snaps since
some versions. The file is used to ship boot assets that are
part of the kernel.

The first PR with support can be found here:
https://github.com/snapcore/snapd/pull/9150

This commit adds support for including a `kernel.yaml` as
meta/kernel.yaml in the snap. This will help the pi-kernel
to ship their DTBs as part of the kernel snap.