storeapi: move http client and auth to http_clients package
Move the existing http clients (UbuntuOneAuthClient and Client)
together with the agent logic into its own http_clients.

Corresponding errors (which still need redoing with a cetralized
strategy around the final request) to this package too.

Additionally, migrate snapcraft.config.Config to this package in the
form of a Abstract Base Class with a specific implementation for the
UbuntuOneAuthClient class.

While migrating config, the "local credentials" implementation was
removed as it was a feature brought in from the click_toolbelt days
and we never really exposed in Snapcraft.

Signed-off-by: Sergio Schvezov <sergio.schvezov@canonical.com>