godeps v1 plugin: set default channel to 1.15/stable
The godeps plugin breaks with 1.16 due to a change of behavior
with go get (go/src/... is no longer populated as the plugin
expects).

As godeps is no longer maintained, it doesn't really make sense
to update the plugin to work with 1.16.  Simply sets the plugin's
default to the last-known working version of go, 1.15.

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>