store: use whoami dashboard endpoint for cli
Add the MVP support for the whoami endpoint to the Store APIs under
Dashboard to serve as a replacement for using the account information
endpoint to serve as the backend to the whoami command.

Caching was removed from the command as the existing caching was too
strict (over caching).

"whoami" command test was rewritten in pure pytest.

Signed-off-by: Sergio Schvezov <sergio.schvezov@canonical.com>