elf: add fallback methods to detect library dependencies (#3514)
libc6 may mismatch with the paths provided to ldd in LD_LIBRARY_PATH.

This may be avoided if invoking the binary directory and using
LD_TRACE_LOADED_OBJECTS, but this method is limited to executable
files.

This updates the library dependency search to try three variants:
(1) use ldd as today.  if ldd fails..
(2) use ldd with LD_PRELOAD to specify libc6 from host. if ldd fails...
(3) use ld trace method. if ldd trace fails...
(4) warn as before.

For #2, libc6 is expected as /lib/<arch-triplet>/libc.so.6.  This likely
will need to be more robust in the future.

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>