yaml: detect and log warn on duplicate keys in snapcraft.yaml
A very primitive check to warn if duplicate keys are found in
the snapcraft.yaml.  At the moment the file is loaded multiple
times which may result in duplicate warnings (ha ha) which we
need to address in future work.  Given that there shouldn't be
duplicates in YAML files, the extra verbosity should be OK.

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>