cli: add support for parallel remote builds
Currently, when spinning up a remote build, snapcraft calculates a hash
of the path for the project being built, and uses that as a unique
identifier for the build. This artificially limits the number of builds
that can be happening for a given project to 1, which really limits the
usefulness of remote builds.

Instead of calculating a hash for the project directory, calculate a
hash that represents the _state_ of the project, such that changing any
file in the working tree (e.g. switching git branches) changes the hash.
This allows multiple builds for different versions of a given project to
be running at the same time.

Of course, this introduces the complication of needing to track such
builds. Expose the concept of a "build ID" to the user and allow them to
specify it for every remote-build command, so they can check the status
of or recover particular builds by providing the relevant build ID.

Signed-off-by: Kyle Fazzari <kyrofa@ubuntu.com>