Autotools Plugin: Fix fatal crash when running autogen.sh or bootstrap
The autotools plugin will attempt to run `autogen.sh` or `bootstrap` scripts if they are present in the part's source. However, the internal representation of these scripts is a `Path` object, not a string, so when passed to `shlex` it throws an exception stating that it was expecting a bytes-like object.

Signed-off-by: Dani Llewellyn <dani@bowlhat.net>