parts: integrate package-repositories
Add apt package repositories to the host system according to the
`package-repositories` entry listed in snapcraft.yaml.

Also add a separate set of spread tests for core22. These can be
removed and integrated to the existing tests when better 22.04
support is added to the testing infrastructure and more options
are implemented in the 7.0 branch.

Note: The pydantic model for package repositories are very similar
to the legacy internal classes used to manipulate deb and ppa
repositories. We should investigate if these can be consolidated,
moving manual validation of repository parameters to the pydantic
model.

Signed-off-by: Claudio Matsuoka <claudio.matsuoka@canonical.com>