gradle v1 plugin: add support for JDK 17
Allow users to choose JDK 17.

JDK 17 is available in bionic:
https://packages.ubuntu.com/search?suite=bionic&searchon=names&keywords=openjdk-17-jdk
Not having this feature causes my snap to both have JRE 11 *and* 17:
https://github.com/EduMIPS64/edumips64/issues/657

In an attempt to not inadvertently have lots of snaps build with JDK 17,
also change the default JDK to be 11, changing the default stated
behavior from auto-picking the last JDK available in the base (which was
not happening anyway) to auto-picking version 11.