extensions: support core22
This adds some variations to the API and changes where things are
processed when compared to the original implementation.

For integration it is rather hard to test until there is an extension
implementation to test with, so the current code only makes sure that
use of apply_extensions does not break the current implementation.

Signed-off-by: Sergio Schvezov <sergio.schvezov@canonical.com>