storeapi: environment variable for auth selection
Auth selection, Ubuntu Login SSO or Candid, is now donw through an
environment variable. This cleans up the internals and simplifies the
API. It also allows us to move away from it when the time comes to maybe
OAUTH.

This code removes support for the --experimental-login option in
commands that used to support it with an error message for the new
mechanism.

Signed-off-by: Sergio Schvezov <sergio.schvezov@canonical.com>