extensions: don't fail during installation
Currently, snapd doesn't connect automatically the plug for
gnome-42-2204, which means that, during installation, all the files
contained there aren't available. One of that files is fc-cache,
used in the 'configuration' hook to update the fonts.

This patch checks wether it is available or not before trying to
execute it, thus avoiding the failure during installation.