linters: add linting infrastructure
Create a basic framework for linting including the `lint` entry in
snapcraft.yaml to disable specific linters. Linters run before
packing a snap.

A specific `lint` command to lint a snap package file will be added
in a future PR. JSON-formatted output is implemented but will only
be used with the `lint` command.

Signed-off-by: Claudio Matsuoka <claudio.matsuoka@canonical.com>