store: improve error message for 401 error
When the store returns a 401, Snapcraft tries to do the right thing to
prompt for a new login, but the error messages where confusing for the
cases it cannot, when this original code was added the concept of a
legacy client was not in place.

A logout is done to enforce using the new client if the credentials are
the configparser based ones.

LP: #1980533

Signed-off-by: Sergio Schvezov <sergio.schvezov@canonical.com>