commands, lifecycle: provide snapcraft try for core22 (#3981)
* commands, lifecycle: provide snapcraft try for core22

This commit adds support for ``snapcraft try`` for core22 snaps. Just
like in the legacy version, this works by creating a ``prime`` directory
on the host that is then mounted into the ``/root/prime`` directory
inside the instance. The lifecycle then runs normally and the project is
primed but not packed.