requirements: update to craft-parts 1.18.0
Use craft-parts' optional stage package tracking only when manifest
is enabled. This allows Snapcraft to run on instances based on tmpfs,
for faster builds and snap development cycles.

This release also adds plugins for SCons, Ant, and Maven, and fixes
directory cleaning. See the Craft-parts project changelog for a
complete list of changes.

Signed-off-by: Claudio Matsuoka <claudio.matsuoka@canonical.com>