spread: update flutter test
The test had sdk bounds that the latest master would not work with (new DART),
while removing the bounds, the pubspec was outdated enough.

Moved to the new flutter template post linux desktop beta announcement. The
downside is that `<app> --help` now requires X (could be solved with xfvb) but
alas, does not print help and exit.

Signed-off-by: Sergio Schvezov <sergio.schvezov@canonical.com>