Parts: Allow to clean all the parts
This MR adds the option of using

    snapcraft clean --keep-environment

to clean all the parts in a single step. This allows to fully
rebuild the snap, but without having to recreate the virtual
machine or the container, like when using "snapcraft clean".
This saves a lot of time.

It's possible to use -k instead of --keep-environment, if
preferred.