fix(clean): non-host cleaning shouldn't load the project
snapcraft try can create a prime/ directory on the host even when
dealing with containers and VM.
The call to the lifecycle.clean method calls the
project_loader.load_config which calls the pluginhandler.load_plugin

This load_plugin method is using sys.path to load plugins.
This is using the host sys.path.
If the host has a python module with the same name as the snapcraft
plugin, it will try to load the module from the host.
The module from the host being something completly unrelated to
snapcraft plugins, the load crashes.

Adding a method only responsible for deleting the "try"
directory avoids this issue