linters: allow "categories" for subfiltering
This commit adds the notion of "categories" for Linter subclasses. For
linters that perform multiple different "types" of linting, this lets
the author ignore specific paths for specific types of linting issues,
with a finer granularity than ignoring the whole linter wholesale.

Currently the only Linter type that supports this is the LibraryLinter,
which provides the "unused-library" and "missing-library" categories to
disable checking for unused or missing libraries.