python v2 plugin: support building parts conforming to PEP 518
Currently Snapcraft only do a `pip install .` operation when the
setup.py file is found in the part's source tree[1], which is no longer
the case with projects adhering to PEP 518[2] as such projects may have
a pyproject.toml file in their source tree instead(e.g. OCRmyPDF[3]).

This patch instructs Snapcraft to run the `pip install .` command when
pyproject.toml is found in the source tree, according to the pip pull
request that implements the support[4] and the Ubuntu package search
results[5] this should be supported since core20.

[1]: [Line 143 of snapcraft/python.py at 0e2cf91 · snapcore/snapcraft · GitHub](https://github.com/snapcore/snapcraft/blob/0e2cf91/snapcraft_legacy/plugins/v2/python.py#L143)
[2]: [PEP 518 – Specifying Minimum Build System Requirements for Python Projects | peps.python.org](https://peps.python.org/pep-0518/)
[3]: [Drop setup.py - use pyproject.toml exclusively · ocrmypdf/OCRmyPDF@16fc520](https://github.com/ocrmypdf/OCRmyPDF/commit/16fc520)
[4]: [Install build dependencies as specified in PEP 518 by takluyver · Pull Request #4144 · pypa/pip](https://github.com/pypa/pip/pull/4144)
[5]: [Ubuntu – Package Search Results -- python3-pip](https://packages.ubuntu.com/search?lang=en&suite=all&exact=1&searchon=names&keywords=python3-pip)

LP: #2008237

Signed-off-by: 林博仁(Buo-ren Lin) <Buo.Ren.Lin@gmail.com>