parts: show the current part lifecycle in the window title
When building a snap with several parts using -v, it's usually
not possible to know which part is being built due to the long
text output.

This patch shows in the window title the part lifecycle that is
being executed.

Although maybe a better way would be to add a method "set_title"
in craft_cli... but that would require two PRs.