tests: provide a fixture with the project's 'main' module (#99)
The idea is this: since we have a few tests that are worth running in
"downstream" projects (those that merge starbase periodically), instead
of having to adapt every test individually in every downstream project
we'll use a fixture (project_main_module) that provides the imported
module object. So in starbase is returns the 'starcraft' module.

The tests are then updated to use this fixture instead of importing the
module directly. Downstream projects therefore only need to update the
fixture's implementation and the tests will (well, should) work.