requirements: bump craft-archives
Version 1.1.3 fixes the issue with package-repositories that declare
architectures. In addition to addressing the wrong parameter order, the
new version also improves the handling of architectures.

The summary of the change is: if the host system is amd64 and the arch
of the package repository is i386, _or_ if the host is arm64 and the
repo armhf, then `dpkg --add-architecture <repo_arch>` will be called.
Otherwise, the command is not called. This lets us preserve the
behavior of enabling i386 in the official repositories when a repo
using i386 is added, with the improvement of _not breaking_ apt update
when an "incompatible" architecture combination is declared.

The new spread tests reflect this. Both run in amd64 and add the
deadsnakes ppa to fetch Python in a foreign arch:

- For i386, only the deadsnakes ppa is necessary, as the architecture
will be added "globally" and Python's dependencies are found in the
default archives;
- For armhf, the ports.ubuntu.com repo is necessary in addition to the
deadsnakes one. This lets the dependencies be found in the ports repo,
_and_ doesn't break the build (this combination is broken in Snapcraft
7.2, 7.3 and 7.4).

Fixes #4289 and #4298