feat!: fail if we know the project is core24
This commit updates the scaffolding to have a different behavior in
cases where we *know* that the existing project (from the yaml file) is
core24-based.

In this cases, we now raise a RuntimeError outlining the name of the
command instead of raising a ClassicFallback error to let the
non-craft-application workflow take over.

With this, Snapcraft runs on core24 projects will fail until the
commands are ported over, while projects for older bases (or commands
that don't use projects, like "init") will still fallback to the
existing behavior.