fix(spread): update tests that manipulate architectures
This reflects a change in behavior in core24:

- Trying to build multiple snaps in a single run is an error;
- Trying to build no snaps at all is also an error.

Some of the subtests in architectures/ are still failing, but that's because
they need other fixes (like SNAPCRAFT_BUILD_FOR and "build-for: all" support).

In addition, the build-for-all test is removed because we decided to _not_
support "build-for: all" for now in core24, because the concept itself of
multiplatform snaps needs to be further discussed.