feat: adopt more fields from the appstream metadata file(part 2)
This is the Part 2 of PR #4647. This adds a new method named `_update_project_links` which is called from the `update_project_metadata` function. It checks if the `project` has that particular metadata and if yes, then it appends the value into it, and else it casts the value onto it from the `extracted_metadata`