fix: use "devel" bases on build plan
For the BuildInfos we want to use "devel" bases, as that is the term that
craft-providers expects to create instances based on the development track.
We still need to use the regular "base" (e.g. "core24") in other places, so
add a customization parameter to get_effective_base() to use either behavior.