feat(plugins): support core24 in the Python plugin
The interpreter for confined core24 snaps is /usr/bin/python3.12. This also
includes some minor rework of the logic: if the snap's confinement is strict
or devmode, then we should definitely know which interpreter to use. This
can make our lives easier in the future as we support more bases.