Activate installed_base_by_architecture metrics
Pursuant to my post on Discourse
(https://forum.snapcraft.io/t/enable-by-architecture-in-the-snapcraft-metrics-cli-tooling/39750),
this commit activates the installed_base_by_architecture and weekly_installed_base_by_architecture
metrics in the snapcraft metrics tool.

The metric was previously introduced according to this PR
https://github.com/canonical/snapcraft.io/issues/1680#issuecomment-688928336

A subsequent issue will be raised against the documentation.