fix: install gpg and dirmngr in core24 builds
This commit mimicks the behavior for core22<= builds where gpg and dirmngr
are installed at runtime for projects with package-repositories. The trigger
for this commit is the fact that the environment where Snapcraft builds in
Launchpad do _not_ have the 'dirmngr' package installed.

Unfortunately the strategy of bundling 'gpg' and 'dirmngr' as stage-packages
in Snapcraft's own snap didn't work because gpg has the expected path to the
'dirmngr' executable hardcoded (always trying to call '/usr/bin/dirmngr').

Fixes #4740