build(docs): fix version in rtd builds (#4864)
Readthedocs builds don't fetch tags when cloning the repo. This is a
problem for us because Snapcraft's version comes from the git tag, so do
a fetch of the tags as a post-checkout step.

Additionally, explicitly install snapcraft itself (via "pip install .")
so that the code that gets the version from git actually runs. Note that
this already works on "latest" because that builds off of "main" where
this fix has already been applied (when moving to canonical-sphinx).