fix: fix Python venv in core24 classic snaps
Changes to virtual environment handling made our default approach - creating a
venv in the build system and then just moving it into the snap - not work for
Python 3.12.

My understanding of the issue is that the "home" key in the "pyvenv.cfg" file
points to a path in the build system, which obviously does not exist at
snap-run-time. This apparently works fine *somehow* in core22 - likely there's
some fallback logic that makes the interpreter correctly find the bundled
Python libraries, but not in core24. So to address this we update this "home"
key to point to its final destination in the snap.

Fixes #4942