remote-build: add "Pending" as a possible status output
Currently, remote-build status goes from "Uploading" to "Stopped".
`BuildState` has a PENDING state that is set but not checked for. Going
from "Uploading" to "Pending" improves status messages significantly.
The output could also say "Needs building" to match the web view, but
"Pending" is more in line with the other single-word outputs.