build(deps): bump craft-parts
This new version fixes a bug on the npm plugin. The bug is caused by the plugin
having state that implicitly needs to be passed from the 'pull' step to the
'build' step. If both are done in a single Snapcraft 'run', then it works fine;
however, running first 'pull' and then 'build' separately trigger the issue -
a failure on the build step.

Ref: https://github.com/canonical/craft-parts/issues/844