docs: improve sidebar for the changelog page
These two facts:

- Changelog entries for a given version can be rather long
- We order the versions "chronologically" instead of "numerically"

... mean that the entries for a given major version can get "hidden" further
down the page if multiple hotfixes have been released in the meantime. This
can cause confusion and make it seem like the version has not been released,
or at least doesn't have changelog entries. To alleviate this, add a table of
contents to the sidebar that links to the subsection for each release.