feat: add Snapcraft-specific Poetry plugin
The plugin has the same behavior and restrictions as the Python plugin with
regards to base-dependent behavior, symlink handling, etc.

Therefore, this common behavior is extracted into a new "python_common" module,
used by both plugins. This approach is also taken for the reference docs - the
requirement of staging a Python interpreter (or not) is the same for both
plugins.

Enabled for core22 and core24.

Fixes #5025