build!: switch from using extras to dependency groups (#330)
* build: switch from using extras to dependency groups

Now that PEP 735 is finalised and we have a standard for dependency
groups, we should move away from using extras to using those.

https://packaging.python.org/en/latest/specifications/dependency-groups/

* build(docs): use uv in readthedocs

This adjusts the readthedocs to use uv, which speeds up the RTD build
time fairly significantly. The old RTD configuration also would not work
now that we've moved from extras to dependency groups.

* ci: fix the pull request template to use make