fix(remotebuild): don't fail when cross-compiling (#5265)
This PR updates Snapcraft to pass `build-on` architectures to Launchpad
rather than the `build-for` architectures. It updates the spread tests
to exercise this.

Benefits:
* Previously, a snap with `build-for: [all]` would build on `amd64`. Now
it builds on one of the `build-on` architectures in the
`snapcraft.yaml`.
* Cross-compiling works with remote builds.

This is possible due to the following constraints:
* Snapcraft only allows one artifact to be remotely built on each
`build-on` architecture.
* Launchpad accepts a list of architectures where to build an artifact
and choose one architecture from that list.

Signed-off-by: Callahan Kovacs <callahan.kovacs@canonical.com>