test: update ua token test (#5322)
This test had never been running because the secret was named
`SNAPCRAFT_UA_TOKEN` in Github instead of `UA_TOKEN`. The test had
silently passed with `exit 0`.

I added a new Github secret with the correctly named UA token and
updated the test.

This follows the same precedence defined in the store tests.

Signed-off-by: Callahan Kovacs <callahan.kovacs@canonical.com>