feat: add init test profile
Add template for spread.yaml and spread extension files to the
"test"  init profile. This was added to a separate profile to
make it more convenient to test this feature without disturbing
the default profile, and to allow test files to be added to
existing projects. After test files stabilize, they can be also
added to the default profile.

Signed-off-by: Claudio Matsuoka <claudio.matsuoka@canonical.com>