fix: add all parts as a dep of extension part (#5395)
Previously with cached builds, there'd always be cases, where
files from the extension's part would not be found as the files are removed
from prime, but the extension's part's state shows it as built.
Adding all the parts as a dependency of the extension ensures that the extension's
part is always rebuilt whenever on part is changed.

Co-authored-by: Callahan <callahan.kovacs@canonical.com>