feat(gnome): Add sdk include directories to C_INCLUDE_PATH
The sdk provides a base /usr/include path that should be checked when an
including from the default paths.

Without this, when a file is including something such as
  #include <xkbcommon/xkbcommon.h>

that is provided by /snap/gnome-46-2404-sdk/current/usr/include, the
compiler won't look for such path at all.

However do this only for core24, because in core22 the clang version does
not handle it properly, breaking flutter builds